Add MySQL subchart to Helm chart for Todo app
This commit introduces a MySQL subchart to the Helm chart under the 'todoapp' directory, enhancing the application's database management capabilities. A new .helmignore file is added to prevent unnecessary files from being packaged. Additionally, the version of the main todoapp Chart.yaml is updated to 1.0.0 to mark this significant enhancement.